% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Indy_leaflet.R
\name{make_Indy_leaflet}
\alias{make_Indy_leaflet}
\title{Make a leaflet basemap of Indianapolis, IN}
\usage{
make_Indy_leaflet()
}
\value{
A leaflet htmlwidget
}
\description{
Make a leaflet basemap of Indianapolis, IN
}
